+++ include/rhbm_gem/data/object/AtomObject.hpp
@@ -22,7 +22,7 @@ class AtomObject
     bool m_is_special_atom;
     int m_serial_id, m_residue_id;
     std::string m_component_id, m_atom_id, m_chain_id, m_indicator;
-    float m_occupancy, m_temperature;
+    double m_occupancy, m_temperature;
     double m_standard_qscore;
     ComponentKey m_component_key;
     AtomKey m_atom_key;
@@ -31,10 +31,10 @@ class AtomObject
     Spot m_spot;
     Structure m_structure;
     ModelObject * m_owner_model;
-    std::array<float, 3> m_position;
-    std::unordered_map<std::string, std::array<float, 3>> m_alternate_position_map;
-    std::unordered_map<std::string, float> m_alternate_occupancy_map;
-    std::unordered_map<std::string, float> m_alternate_temperature_map;
+    std::array<double, 3> m_position;
+    std::unordered_map<std::string, std::array<double, 3>> m_alternate_position_map;
+    std::unordered_map<std::string, double> m_alternate_occupancy_map;
+    std::unordered_map<std::string, double> m_alternate_temperature_map;
 
 public:
     AtomObject();
@@ -50,18 +50,18 @@ public:
     void SetAtomKey(AtomKey value) { m_atom_key = value; }
     void SetChainID(const std::string & value) { m_chain_id = value; }
     void SetIndicator(const std::string & value) { m_indicator = value; }
-    void SetOccupancy(float value) { m_occupancy = value; }
-    void SetTemperature(float value) { m_temperature = value; }
+    void SetOccupancy(double value) { m_occupancy = value; }
+    void SetTemperature(double value) { m_temperature = value; }
     void SetStandardQScore(double value) { m_standard_qscore = value; }
     void SetResidue(Residue value) { m_residue = value; }
     void SetElement(Element value) { m_element = value; }
     void SetSpot(Spot value) { m_spot = value; }
     void SetStructure(Structure value) { m_structure = value; }
-    void SetPosition(float x, float y, float z);
-    void SetPosition(const std::array<float, 3> & value);
-    void AddAlternatePosition(const std::string & indicator, const std::array<float, 3> & value);
-    void AddAlternateOccupancy(const std::string & indicator, float value);
-    void AddAlternateTemperature(const std::string & indicator, float value);
+    void SetPosition(double x, double y, double z);
+    void SetPosition(const std::array<double, 3> & value);
+    void AddAlternatePosition(const std::string & indicator, const std::array<double, 3> & value);
+    void AddAlternateOccupancy(const std::string & indicator, double value);
+    void AddAlternateTemperature(const std::string & indicator, double value);
     std::string GetInfo() const;
     ComponentKey GetComponentKey() const { return m_component_key; }
     AtomKey GetAtomKey() const { return m_atom_key; }
@@ -78,14 +78,14 @@ public:
     std::string GetAtomID() const { return m_atom_id; }
     std::string GetChainID() const { return m_chain_id; }
     std::string GetIndicator() const { return m_indicator; }
-    float GetOccupancy() const { return m_occupancy; }
-    float GetTemperature() const { return m_temperature; }
+    double GetOccupancy() const { return m_occupancy; }
+    double GetTemperature() const { return m_temperature; }
     double GetStandardQScore() const { return m_standard_qscore; }
-    std::array<float, 3> GetPosition() const { return m_position; }
-    const std::array<float, 3> & GetPositionRef() const { return m_position; }
-    const std::unordered_map<std::string, std::array<float, 3>> & GetAlternatePositions() const;
-    const std::unordered_map<std::string, float> & GetAlternateOccupancies() const;
-    const std::unordered_map<std::string, float> & GetAlternateTemperatures() const;
+    std::array<double, 3> GetPosition() const { return m_position; }
+    const std::array<double, 3> & GetPositionRef() const { return m_position; }
+    const std::unordered_map<std::string, std::array<double, 3>> & GetAlternatePositions() const;
+    const std::unordered_map<std::string, double> & GetAlternateOccupancies() const;
+    const std::unordered_map<std::string, double> & GetAlternateTemperatures() const;
     std::vector<AtomObject *> FindNeighborAtoms(double radius = 2.5, bool include_self = false) const;
 
 private:
