+++ include/rhbm_gem/utils/math/GridSampler.hpp
@@ -5,8 +5,8 @@
 class GridSampler
 {
     unsigned int m_sampling_size;
-    float m_window_size;
-    std::array<float, 3> m_reference_u_vector;
+    double m_window_size;
+    std::array<double, 3> m_reference_u_vector;
 
 public:
     GridSampler();
@@ -14,12 +14,12 @@ public:
 
     void Print() const;
     SamplingPointList GenerateSamplingPoints(
-        const std::array<float, 3> & reference_position,
-        const std::array<float, 3> & plane_normal) const;
+        const std::array<double, 3> & reference_position,
+        const std::array<double, 3> & plane_normal) const;
     unsigned int GetGridResolution() const { return m_sampling_size; }
     unsigned int GetPointCount() const { return m_sampling_size * m_sampling_size; }
     void SetGridResolution(unsigned int value) { m_sampling_size = value; }
-    void SetWindowSize(float value) { m_window_size = value; }
-    void SetReferenceUVector(const std::array<float, 3> & value) { m_reference_u_vector = value; }
+    void SetWindowSize(double value) { m_window_size = value; }
+    void SetReferenceUVector(const std::array<double, 3> & value) { m_reference_u_vector = value; }
 
 };
