+++ include/rhbm_gem/utils/math/SphereSampler.hpp
@@ -7,11 +7,11 @@
 namespace rhbm_gem::sphere_sampler {
 
 SamplingPointList GenerateSamplingPointList(
-    const std::array<float, 3> & center_position,
+    const std::array<double, 3> & center_position,
     SphereSamplingMethod method);
 
-SamplingPointList GenerateVolumeUniformRandom(const std::array<float, 3> & center_position);
-SamplingPointList GenerateRadiusUniformRandom(const std::array<float, 3> & center_position);
-SamplingPointList GenerateFibonacciDeterministic(const std::array<float, 3> & center_position);
+SamplingPointList GenerateVolumeUniformRandom(const std::array<double, 3> & center_position);
+SamplingPointList GenerateRadiusUniformRandom(const std::array<double, 3> & center_position);
+SamplingPointList GenerateFibonacciDeterministic(const std::array<double, 3> & center_position);
 
 } // namespace rhbm_gem::sphere_sampler
